thermal-engine: use bms sensor for charging throttling
17th sensor get's way above 70C after 7-10 minutes of
CPU stress test and above 50C in normal usage. This kills charging
entirely. We do not want that because actually it does not
indicate that we are overheating. What we really need to check is
the battery temperature.
This change ensures that we have at least 400-500mA charging
even under heavy usage.

bms and battery sensors show the same temperature.

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: Davide Garberi <dade.garberi@gmail.com>